requirements.txt
Django
psycopg2-binary
#psycopg2
gunicorn
asgiref
autopep8
#backports.zoneinfo;python_version<"3.9"
pycodestyle
sqlparse
toml
django-admin-interface
django-rest-framework
django-import-export
django-crispy-forms
crispy-bootstrap5
django-bootstrap-datepicker-plus
#matplotlib
mpl_finance
yfinance    
ccxt
mpld3
slack_sdk
slackclient
requests
python-dotenv
celery[redis]
django-celery-beat
netmiko